Rework combobox for keyboard navigation and simpler chips
Replace cmdk with a hand-rolled listbox so multi-select can be driven
entirely from the keyboard. cmdk scoped its arrow/Enter navigation to
its own DOM subtree, which the portaled popover broke, leaving the
trigger input unable to move through or pick options.

- Keyboard navigation now works in both single and multi-select: arrows
  move the active option, Enter selects, Home/End jump, Escape closes,
  with the active option exposed via aria-activedescendant and scrolled
  into view.
- Filtering is pure case-insensitive substring matching over label and
  value for both modes, replacing cmdk's fuzzy filter.
- Selected chips grow and wrap by default so every selection stays
  visible; maxVisibleChips collapses the remainder into a "+N" summary
  for a compact fixed-row layout. This removes the measured-overflow
  machinery (ResizeObserver, offscreen measurement) and the layout bugs
  it caused, including the search input wrapping under the chips.
- Split the component into focused files (context, root, trigger,
  content, convenience) so no single file carries the whole family.

Refs #306
