+++ libs/shared/react/ui/src/components/combobox/combobox.tsx
@@ -1,818 +1,11 @@
 'use client';
 
-import * as React from 'react';
-import {cn} from '#utils/cn.js';
-import {
-  Command,
-  CommandEmpty,
-  CommandGroup,
-  CommandInput,
-  type CommandInputProps,
-  CommandItem,
-  CommandList,
-  CommandTrigger,
-  type CommandTriggerProps,
-  commandTriggerVariants,
-} from '../command/index.js';
-import {Icon} from '../icon/index.js';
-import {Popover, PopoverContent, PopoverTrigger} from '../popover/index.js';
-import {ScrollArea} from '../scroll-area/index.js';
-import {
-  type ComboboxOption,
-  clearMultiComboboxValues,
-  partitionComboboxChipsByCount,
-  partitionComboboxChipsByWidth,
-  removeMultiComboboxValue,
-  resolveComboboxLabel,
-  toggleMultiComboboxValue,
-  toggleSingleComboboxValue,
-} from './combobox-state.js';
-
-export type {ComboboxOption};
-
-type ComboboxRootBaseProps = {
-  options: ComboboxOption[];
-  children: React.ReactNode;
-  disabled?: boolean;
-  isLoading?: boolean;
-  maxVisibleChips?: number | 'auto';
-};
-
-type SingleControlledComboboxRootProps = ComboboxRootBaseProps & {
-  multiple?: false;
-  value: string;
-  defaultValue?: never;
-  onValueChange?: (value: string) => void;
-};
-
-type SingleUncontrolledComboboxRootProps = ComboboxRootBaseProps & {
-  multiple?: false;
-  value?: never;
-  defaultValue?: string;
-  onValueChange?: (value: string) => void;
-};
-
-type MultiControlledComboboxRootProps = ComboboxRootBaseProps & {
-  multiple: true;
-  value: string[];
-  defaultValue?: never;
-  onValueChange?: (value: string[]) => void;
-};
-
-type MultiUncontrolledComboboxRootProps = ComboboxRootBaseProps & {
-  multiple: true;
-  value?: never;
-  defaultValue?: string[];
-  onValueChange?: (value: string[]) => void;
-};
-
-export type ComboboxRootProps =
-  | SingleControlledComboboxRootProps
-  | SingleUncontrolledComboboxRootProps
-  | MultiControlledComboboxRootProps
-  | MultiUncontrolledComboboxRootProps;
-
-type ComboboxContextValue = {
-  options: ComboboxOption[];
-  multiple: boolean;
-  disabled: boolean;
-  isLoading: boolean;
-  maxVisibleChips: number | 'auto';
-  listId: string;
-  open: boolean;
-  setOpen: (open: boolean) => void;
-  searchValue: string;
-  setSearchValue: (value: string) => void;
-  selectedValue: string;
-  selectedValues: string[];
-  getLabel: (value: string) => string;
-  isSelected: (value: string) => boolean;
-  selectValue: (value: string) => void;
-  removeValue: (value: string) => void;
-  removeLastValue: () => void;
-  clearValues: () => void;
-};
-
-const ComboboxContext = React.createContext<ComboboxContextValue | null>(null);
-
-function useComboboxContext() {
-  const context = React.useContext(ComboboxContext);
-  if (!context) {
-    throw new Error('Combobox components must be used within a ComboboxRoot.');
-  }
-  return context;
-}
-
-function ComboboxRoot(props: ComboboxRootProps) {
-  const {options, children, disabled = false, isLoading = false, maxVisibleChips = 'auto'} = props;
-  const multiple = props.multiple === true;
-  const propsRef = React.useRef(props);
-  propsRef.current = props;
-  const listId = React.useId();
-  const [open, setOpen] = React.useState(false);
-  const [searchValue, setSearchValue] = React.useState('');
-  const [internalSingleValue, setInternalSingleValue] = React.useState(
-    !multiple ? (props.defaultValue ?? '') : '',
-  );
-  const [internalMultiValue, setInternalMultiValue] = React.useState(
-    multiple ? (props.defaultValue ?? []) : [],
-  );
-  const handleOpenChange = React.useCallback(
-    (nextOpen: boolean) => {
-      if (!disabled) {
-        setOpen(nextOpen);
-      }
-    },
-    [disabled],
-  );
-
-  const selectedValue = multiple
-    ? ''
-    : ((props as SingleControlledComboboxRootProps | SingleUncontrolledComboboxRootProps).value ??
-      internalSingleValue);
-  const controlledMultiValue = multiple
-    ? (props as MultiControlledComboboxRootProps | MultiUncontrolledComboboxRootProps).value
-    : undefined;
-  const selectedValues = React.useMemo<string[]>(
-    () =>
-      multiple
-        ? (controlledMultiValue ?? internalMultiValue)
-        : selectedValue
-          ? [selectedValue]
-          : [],
-    [multiple, controlledMultiValue, internalMultiValue, selectedValue],
-  );
-
-  const updateSingleValue = React.useCallback(
-    (nextValue: string) => {
-      const singleProps = propsRef.current as
-        | SingleControlledComboboxRootProps
-        | SingleUncontrolledComboboxRootProps;
-
-      if (!multiple && singleProps.value === undefined) {
-        setInternalSingleValue(nextValue);
-      }
-      if (!multiple) {
-        singleProps.onValueChange?.(nextValue);
-      }
-    },
-    [multiple],
-  );
-
-  const updateMultiValue = React.useCallback(
-    (nextValues: string[]) => {
-      const multiProps = propsRef.current as
-        | MultiControlledComboboxRootProps
-        | MultiUncontrolledComboboxRootProps;
-
-      if (multiple && multiProps.value === undefined) {
-        setInternalMultiValue(nextValues);
-      }
-      if (multiple) {
-        multiProps.onValueChange?.(nextValues);
-      }
-    },
-    [multiple],
-  );
-
-  const getLabel = React.useCallback(
-    (value: string) => resolveComboboxLabel(options, value),
-    [options],
-  );
-
-  const isSelected = React.useCallback(
-    (value: string) => (multiple ? selectedValues.includes(value) : selectedValue === value),
-    [multiple, selectedValue, selectedValues],
-  );
-
-  const selectValue = React.useCallback(
-    (value: string) => {
-      if (disabled) {
-        return;
-      }
-
-      if (multiple) {
-        updateMultiValue(toggleMultiComboboxValue(selectedValues, value));
-        setSearchValue('');
-        return;
-      }
-
-      updateSingleValue(toggleSingleComboboxValue(selectedValue, value));
-      setSearchValue('');
-      setOpen(false);
-    },
-    [disabled, multiple, selectedValue, selectedValues, updateMultiValue, updateSingleValue],
-  );
-
-  const removeValue = React.useCallback(
-    (value: string) => {
-      if (disabled || !multiple) {
-        return;
-      }
-      updateMultiValue(removeMultiComboboxValue(selectedValues, value));
-    },
-    [disabled, multiple, selectedValues, updateMultiValue],
-  );
-
-  const removeLastValue = React.useCallback(() => {
-    if (disabled || !multiple || selectedValues.length === 0) {
-      return;
-    }
-    updateMultiValue(selectedValues.slice(0, -1));
-  }, [disabled, multiple, selectedValues, updateMultiValue]);
-
-  const clearValues = React.useCallback(() => {
-    if (disabled || !multiple) {
-      return;
-    }
-    updateMultiValue(clearMultiComboboxValues());
-  }, [disabled, multiple, updateMultiValue]);
-
-  const contextValue = React.useMemo<ComboboxContextValue>(
-    () => ({
-      options,
-      multiple,
-      disabled,
-      isLoading,
-      maxVisibleChips,
-      listId,
-      open,
-      setOpen,
-      searchValue,
-      setSearchValue,
-      selectedValue,
-      selectedValues,
-      getLabel,
-      isSelected,
-      selectValue,
-      removeValue,
-      removeLastValue,
-      clearValues,
-    }),
-    [
-      options,
-      multiple,
-      disabled,
-      isLoading,
-      maxVisibleChips,
-      listId,
-      open,
-      searchValue,
-      selectedValue,
-      selectedValues,
-      getLabel,
-      isSelected,
-      selectValue,
-      removeValue,
-      removeLastValue,
-      clearValues,
-    ],
-  );
-
-  return (
-    <ComboboxContext.Provider value={contextValue}>
-      <Popover open={open} onOpenChange={handleOpenChange}>
-        {children}
-      </Popover>
-    </ComboboxContext.Provider>
-  );
-}
-
-export type ComboboxTriggerProps = Omit<CommandTriggerProps, 'children' | 'placeholder'> & {
-  children?: React.ReactNode;
-  placeholder?: string;
-};
-
-function ComboboxTrigger({
-  children,
-  placeholder = 'Select option...',
-  className,
-  variant,
-  size,
-  disabled,
-  onClick,
-  onKeyDown,
-  id,
-  'aria-label': ariaLabel,
-  'aria-labelledby': ariaLabelledby,
-  'aria-describedby': ariaDescribedby,
-  ...props
-}: ComboboxTriggerProps) {
-  const context = useComboboxContext();
-  const isDisabled = disabled || context.disabled;
-  // Route the label/name onto the labelable widget: the <button> for single-select,
-  // the inline <input> for multi-select. The multi wrapper is a presentational <div>,
-  // which `htmlFor` cannot name, so the input must carry id/aria-* instead.
-  const labelProps = {
-    id,
-    'aria-label': ariaLabel,
-    'aria-labelledby': ariaLabelledby,
-    'aria-describedby': ariaDescribedby,
-  };
-
-  if (!context.multiple) {
-    const triggerContent = children ?? (context.selectedValue ? <ComboboxValue /> : undefined);
-
-    return (
-      <PopoverTrigger asChild>
-        <CommandTrigger
-          variant={variant}
-          size={size}
-          placeholder={placeholder}
-          className={className}
-          disabled={isDisabled}
-          isLoading={context.isLoading}
-          onClick={onClick}
-          onKeyDown={onKeyDown}
-          {...labelProps}
-          {...props}
-        >
-          {triggerContent}
-        </CommandTrigger>
-      </PopoverTrigger>
-    );
-  }
-
-  const triggerContent = children ?? (
-    <ComboboxValue placeholder={placeholder} inputProps={labelProps} />
-  );
-
-  return (
-    <PopoverTrigger asChild>
-      <div
-        data-slot="combobox-trigger"
-        data-disabled={isDisabled || undefined}
-        className={cn(
-          commandTriggerVariants({variant, size}),
-          size === 'small' ? 'min-h-28' : 'min-h-32',
-          'h-auto cursor-text items-start py-4',
-          'data-disabled:pointer-events-none data-disabled:cursor-not-allowed data-disabled:bg-background-neutral-disabled data-disabled:shadow-none data-disabled:text-foreground-neutral-disabled',
-          className,
-        )}
-        {...(props as Omit<React.ComponentProps<'div'>, 'onClick' | 'onKeyDown'>)}
-      >
-        <div className="flex min-w-0 flex-1 flex-wrap items-center gap-4 text-left">
-          {triggerContent}
-        </div>
-        <div className="mt-2 flex shrink-0 items-center gap-4">
-          {context.selectedValues.length > 1 && (
-            <button
-              type="button"
-              aria-label="Clear selected options"
-              disabled={isDisabled}
-              className={cn(
-                'inline-flex size-16 shrink-0 items-center justify-center rounded-4',
-                'text-foreground-neutral-muted transition-colors hover:text-foreground-neutral-base',
-                'focus-visible:shadow-border-interactive-with-active outline-none',
-                'disabled:pointer-events-none disabled:text-foreground-neutral-disabled',
-              )}
-              onClick={(event) => {
-                event.preventDefault();
-                event.stopPropagation();
-                context.clearValues();
-              }}
-            >
-              <Icon name="closeLine" className="size-14" />
-            </button>
-          )}
-          {context.isLoading ? (
-            <Icon name="spinner" className="size-16 text-foreground-neutral-base" />
-          ) : (
-            <Icon name="expandUpDownLine" className="size-16 text-foreground-neutral-muted" />
-          )}
-        </div>
-      </div>
-    </PopoverTrigger>
-  );
-}
-
-type ComboboxValueProps = {
-  placeholder?: string;
-  inputProps?: ComboboxChipsInputProps;
-};
-
-function ComboboxValue({placeholder = 'Select option...', inputProps}: ComboboxValueProps) {
-  const context = useComboboxContext();
-
-  if (context.multiple) {
-    return (
-      <>
-        <ComboboxChips />
-        <ComboboxChipsInput placeholder={placeholder} {...inputProps} />
-      </>
-    );
-  }
-
-  if (!context.selectedValue) {
-    return null;
-  }
-
-  return context.getLabel(context.selectedValue);
-}
-
-type ComboboxContentProps = React.ComponentProps<typeof PopoverContent>;
-
-function ComboboxContent({
-  className,
-  align = 'start',
-  sideOffset = 4,
-  children,
-  ...props
-}: ComboboxContentProps) {
-  const context = useComboboxContext();
-  return (
-    <PopoverContent
-      className={cn('w-(--radix-popover-trigger-width) p-0', className)}
-      align={align}
-      sideOffset={sideOffset}
-      // Multi-select types in the trigger's inline combobox input, so keep focus there
-      // instead of letting the popover steal it on open.
-      onOpenAutoFocus={context.multiple ? (event) => event.preventDefault() : undefined}
-      onWheel={(e) => e.stopPropagation()}
-      onTouchStart={(e) => e.stopPropagation()}
-      onTouchMove={(e) => e.stopPropagation()}
-      {...props}
-    >
-      <Command shouldFilter={!context.multiple}>{children}</Command>
-    </PopoverContent>
-  );
-}
-
-function ComboboxInput(props: CommandInputProps) {
-  const context = useComboboxContext();
-
-  return (
-    <CommandInput
-      value={context.searchValue}
-      onValueChange={context.setSearchValue}
-      disabled={context.disabled}
-      {...props}
-    />
-  );
-}
-
-type ComboboxEmptyProps = React.ComponentProps<typeof CommandEmpty>;
-
-function ComboboxEmpty(props: ComboboxEmptyProps) {
-  return <CommandEmpty {...props} />;
-}
-
-type ComboboxListProps = Omit<React.ComponentProps<typeof CommandList>, 'children'> & {
-  children?: React.ReactNode;
-  emptyState?: React.ReactNode;
-};
-
-function ComboboxList({
-  children,
-  emptyState = 'No option found.',
-  className,
-  ...props
-}: ComboboxListProps) {
-  const context = useComboboxContext();
-  // Single-select delegates filtering to cmdk (fuzzy + ranked) via shouldFilter.
-  // Multi-select drives its own substring filter because its search input lives in
-  // the trigger, outside the cmdk Command that owns single-select filtering.
-  const visibleOptions = React.useMemo(() => {
-    if (!context.multiple) {
-      return context.options;
-    }
-
-    const normalizedSearch = context.searchValue.trim().toLowerCase();
-    if (!normalizedSearch) {
-      return context.options;
-    }
-
-    return context.options.filter(
-      (option) =>
-        option.label.toLowerCase().includes(normalizedSearch) ||
-        option.value.toLowerCase().includes(normalizedSearch),
-    );
-  }, [context.multiple, context.options, context.searchValue]);
-
-  return (
-    <ScrollArea>
-      <CommandList
-        id={context.listId}
-        aria-multiselectable={context.multiple || undefined}
-        className={cn('max-h-300', className)}
-        {...props}
-      >
-        {children ?? (
-          <>
-            <ComboboxEmpty>{emptyState}</ComboboxEmpty>
-            <CommandGroup>
-              {visibleOptions.map((option) => (
-                <ComboboxItem key={option.value} option={option} />
-              ))}
-            </CommandGroup>
-          </>
-        )}
-      </CommandList>
-    </ScrollArea>
-  );
-}
-
-type ComboboxItemProps = Omit<React.ComponentProps<typeof CommandItem>, 'value' | 'onSelect'> & {
-  option?: ComboboxOption;
-  value?: string;
-  onSelect?: (value: string) => void;
-};
-
-function ComboboxItem({
-  option,
-  value,
-  children,
-  className,
-  disabled,
-  onSelect,
-  ...props
-}: ComboboxItemProps) {
-  const context = useComboboxContext();
-  const itemValue = option?.value ?? value;
-
-  if (!itemValue) {
-    throw new Error('ComboboxItem requires either an option or a value.');
-  }
-
-  const label =
-    option?.label ?? (typeof children === 'string' ? children : context.getLabel(itemValue));
-  const selected = context.isSelected(itemValue);
-  const isDisabled = disabled || context.disabled;
-
-  return (
-    <CommandItem
-      value={itemValue}
-      keywords={[label]}
-      disabled={isDisabled}
-      aria-checked={context.multiple ? selected : undefined}
-      data-state={selected ? 'checked' : 'unchecked'}
-      className={className}
-      onSelect={() => {
-        if (isDisabled) {
-          return;
-        }
-        context.selectValue(itemValue);
-        onSelect?.(itemValue);
-      }}
-      {...props}
-    >
-      <Icon name="check" className={cn('size-16 mr-8', selected ? 'opacity-100' : 'opacity-0')} />
-      {children ?? label}
-    </CommandItem>
-  );
-}
-
-const chipClassName = cn(
-  'inline-flex h-20 max-w-160 items-center gap-4 rounded-4 border border-tag-neutral-border',
-  'bg-tag-neutral-bg px-6 text-xs font-medium leading-20 text-tag-neutral-text',
-);
-
-const chipTrailingClassName = 'inline-flex size-12 shrink-0 items-center justify-center rounded-2';
-
-const overflowBadgeClassName = cn(chipClassName, 'max-w-none shrink-0');
-
-type ComboboxChipShapeProps = React.ComponentProps<'span'> & {
-  label: string;
-  trailing: React.ReactNode;
-};
-
-// Shared chip geometry. The visible chip and the offscreen measurement clone both
-// render through this so their widths can never drift apart and break the overflow
-// partition, which measures the clone to decide how many real chips fit.
-const ComboboxChipShape = React.forwardRef<HTMLSpanElement, ComboboxChipShapeProps>(
-  ({label, trailing, className, ...props}, ref) => (
-    <span ref={ref} className={cn(chipClassName, className)} {...props}>
-      <span className="truncate">{label}</span>
-      {trailing}
-    </span>
-  ),
-);
-ComboboxChipShape.displayName = 'ComboboxChipShape';
-
-type ComboboxChipsProps = React.ComponentProps<'div'>;
-
-function ComboboxChips({className, ...props}: ComboboxChipsProps) {
-  const context = useComboboxContext();
-  const {containerRef, measureNodes, overflowRef, visibleValues, hiddenCount} =
-    useVisibleComboboxChips();
-
-  if (context.selectedValues.length === 0) {
-    return null;
-  }
-
-  return (
-    <div
-      ref={containerRef}
-      data-slot="combobox-chips"
-      className={cn('relative flex min-w-0 flex-wrap items-center gap-4', className)}
-      {...props}
-    >
-      {visibleValues.map((value) => (
-        <ComboboxChip key={value} value={value} />
-      ))}
-      {hiddenCount > 0 && (
-        <span
-          role="img"
-          className={overflowBadgeClassName}
-          aria-label={`${hiddenCount} more selected`}
-        >
-          +{hiddenCount}
-        </span>
-      )}
-      {context.maxVisibleChips === 'auto' && (
-        <div
-          aria-hidden="true"
-          className="pointer-events-none absolute -left-[9999px] top-0 flex items-center gap-4 opacity-0"
-        >
-          {context.selectedValues.map((value) => (
-            <ComboboxChipShape
-              key={value}
-              ref={(node) => {
-                measureNodes.current.set(value, node);
-              }}
-              label={context.getLabel(value)}
-              trailing={
-                <span className={chipTrailingClassName}>
-                  <Icon name="closeLine" className="size-12" />
-                </span>
-              }
-            />
-          ))}
-          <span ref={overflowRef} className={overflowBadgeClassName}>
-            +{context.selectedValues.length}
-          </span>
-        </div>
-      )}
-    </div>
-  );
-}
-
-type ComboboxChipProps = React.ComponentProps<'span'> & {
-  value: string;
-};
-
-function ComboboxChip({value, className, ...props}: ComboboxChipProps) {
-  const context = useComboboxContext();
-  const label = context.getLabel(value);
-
-  return (
-    <ComboboxChipShape
-      data-slot="combobox-chip"
-      label={label}
-      className={className}
-      trailing={
-        <button
-          type="button"
-          aria-label={`Remove ${label}`}
-          disabled={context.disabled}
-          className={cn(
-            chipTrailingClassName,
-            'text-tag-neutral-icon transition-colors hover:opacity-70',
-            'focus-visible:shadow-border-interactive-with-active outline-none',
-            'disabled:pointer-events-none disabled:text-foreground-neutral-disabled',
-          )}
-          onClick={(event) => {
-            event.preventDefault();
-            event.stopPropagation();
-            context.removeValue(value);
-          }}
-        >
-          <Icon name="closeLine" className="size-12" />
-        </button>
-      }
-      {...props}
-    />
-  );
-}
-
-type ComboboxChipsInputProps = Omit<React.ComponentProps<'input'>, 'value' | 'onChange'>;
-
-function ComboboxChipsInput({
-  className,
-  placeholder,
-  onKeyDown,
-  ...props
-}: ComboboxChipsInputProps) {
-  const context = useComboboxContext();
-
-  return (
-    <input
-      role="combobox"
-      aria-expanded={context.open}
-      aria-controls={context.listId}
-      aria-haspopup="listbox"
-      aria-autocomplete="list"
-      value={context.searchValue}
-      placeholder={context.selectedValues.length === 0 ? placeholder : undefined}
-      disabled={context.disabled}
-      className={cn(
-        'min-w-80 flex-1 bg-transparent text-sm leading-20 outline-none',
-        'placeholder:text-foreground-neutral-muted',
-        'disabled:cursor-not-allowed disabled:text-foreground-neutral-disabled',
-        className,
-      )}
-      onChange={(event) => {
-        context.setSearchValue(event.target.value);
-        context.setOpen(true);
-      }}
-      onKeyDown={(event) => {
-        onKeyDown?.(event);
-        if (event.defaultPrevented) {
-          return;
-        }
-
-        if (event.key === 'Backspace' && context.searchValue === '') {
-          context.removeLastValue();
-        }
-
-        if (event.key === 'Escape') {
-          context.setOpen(false);
-        }
-      }}
-      {...props}
-    />
-  );
-}
-
-function useVisibleComboboxChips() {
-  const context = useComboboxContext();
-  const containerRef = React.useRef<HTMLDivElement | null>(null);
-  const measureNodes = React.useRef(new Map<string, HTMLSpanElement | null>());
-  const overflowRef = React.useRef<HTMLSpanElement | null>(null);
-  const [availableWidth, setAvailableWidth] = React.useState(0);
-  const [measureVersion, setMeasureVersion] = React.useState(0);
-  const selectedValuesKey = context.selectedValues.join('\u0000');
-
-  React.useLayoutEffect(() => {
-    if (context.maxVisibleChips !== 'auto') {
-      return;
-    }
-
-    const container = containerRef.current;
-    if (!container) {
-      return;
-    }
-
-    const measure = () => {
-      setAvailableWidth(container.getBoundingClientRect().width);
-      setMeasureVersion((version) => version + 1);
-    };
-
-    measure();
-
-    if (typeof ResizeObserver === 'undefined') {
-      return;
-    }
-
-    const observer = new ResizeObserver(measure);
-    observer.observe(container);
-
-    return () => observer.disconnect();
-  }, [context.maxVisibleChips]);
-
-  React.useLayoutEffect(() => {
-    void selectedValuesKey;
-    setMeasureVersion((version) => version + 1);
-  }, [selectedValuesKey]);
-
-  const partition = React.useMemo(() => {
-    void measureVersion;
-
-    if (typeof context.maxVisibleChips === 'number') {
-      return partitionComboboxChipsByCount(context.selectedValues, context.maxVisibleChips);
-    }
-
-    const overflowChipWidth = overflowRef.current?.getBoundingClientRect().width;
-    if (!overflowChipWidth || !availableWidth) {
-      return partitionComboboxChipsByCount(context.selectedValues, 2);
-    }
-
-    const valueWidths = new Map<string, number>();
-    for (const value of context.selectedValues) {
-      const width = measureNodes.current.get(value)?.getBoundingClientRect().width;
-      if (!width) {
-        return partitionComboboxChipsByCount(context.selectedValues, 2);
-      }
-      valueWidths.set(value, width);
-    }
-
-    return partitionComboboxChipsByWidth({
-      values: context.selectedValues,
-      valueWidths,
-      availableWidth,
-      overflowChipWidth,
-      gapWidth: 4,
-    });
-  }, [availableWidth, context.maxVisibleChips, context.selectedValues, measureVersion]);
-
-  return {
-    containerRef,
-    measureNodes,
-    overflowRef,
-    visibleValues: partition.visibleValues,
-    hiddenCount: partition.hiddenCount,
-  };
-}
+import type * as React from 'react';
+import type {CommandTriggerProps} from '../command/index.js';
+import {ComboboxContent, ComboboxInput, ComboboxList} from './combobox-content.js';
+import {ComboboxRoot, type ComboboxRootProps} from './combobox-root.js';
+import type {ComboboxOption} from './combobox-state.js';
+import {ComboboxTrigger} from './combobox-trigger.js';
 
 type ComboboxTriggerPassthroughProps = Omit<
   CommandTriggerProps,
@@ -829,7 +22,7 @@ type ComboboxBaseProps = ComboboxTriggerPassthroughProps & {
   align?: 'start' | 'center' | 'end';
   sideOffset?: number;
   isLoading?: boolean;
-  maxVisibleChips?: number | 'auto';
+  maxVisibleChips?: number;
 };
 
 type SingleControlledComboboxProps = ComboboxBaseProps & {
@@ -883,7 +76,7 @@ export function Combobox({
   size,
   isLoading = false,
   disabled = false,
-  maxVisibleChips = 'auto',
+  maxVisibleChips,
   ...triggerProps
 }: ComboboxProps) {
   const rootProps = {
@@ -914,17 +107,3 @@ export function Combobox({
     </ComboboxRoot>
   );
 }
-
-export {
-  ComboboxChip,
-  ComboboxChips,
-  ComboboxChipsInput,
-  ComboboxContent,
-  ComboboxEmpty,
-  ComboboxInput,
-  ComboboxItem,
-  ComboboxList,
-  ComboboxRoot,
-  ComboboxTrigger,
-  ComboboxValue,
-};
