+++ libs/shared/react/ui/src/components/combobox/index.ts
@@ -1,1 +1,18 @@
-export * from './combobox.js';
+export {Combobox, type ComboboxProps} from './combobox.js';
+export {
+  ComboboxContent,
+  ComboboxEmpty,
+  ComboboxInput,
+  ComboboxItem,
+  ComboboxList,
+} from './combobox-content.js';
+export {ComboboxRoot, type ComboboxRootProps} from './combobox-root.js';
+export type {ComboboxOption} from './combobox-state.js';
+export {
+  ComboboxChip,
+  ComboboxChips,
+  ComboboxChipsInput,
+  ComboboxTrigger,
+  type ComboboxTriggerProps,
+  ComboboxValue,
+} from './combobox-trigger.js';
